Move api_from_message and api_to_message from Channel to utils
