+++ junebug/tests/test_channel.py
@@ -3,6 +3,7 @@ from twisted.internet.defer import inlineCallbacks
 from vumi.message import TransportUserMessage
 from vumi.transports.telnet import TelnetServerTransport
 
+from junebug.utils import api_from_message
 from junebug.workers import MessageForwardingWorker
 from junebug.channel import (
     Channel, ChannelNotFound, InvalidChannelType, MessageNotFound)
@@ -275,7 +276,7 @@ class TestChannel(JunebugTestBase):
             })
 
         expected = in_msg.reply(content='testcontent')
-        expected = conjoin(Channel.api_from_message(expected), {
+        expected = conjoin(api_from_message(expected), {
             'timestamp': msg['timestamp'],
             'message_id': msg['message_id']
         })
@@ -284,7 +285,7 @@ class TestChannel(JunebugTestBase):
 
         [dispatched] = self.get_dispatched_messages('channel-id.outbound')
         self.assertEqual(msg['message_id'], dispatched['message_id'])
-        self.assertEqual(Channel.api_from_message(dispatched), expected)
+        self.assertEqual(api_from_message(dispatched), expected)
 
     @inlineCallbacks
     def test_send_reply_message_inbound_not_found(self):
@@ -298,67 +299,3 @@ class TestChannel(JunebugTestBase):
                 'reply_to': 'i-do-not-exist',
                 'content': 'testcontent',
             }), MessageNotFound)
-
-    @inlineCallbacks
-    def test_api_from_message(self):
-        '''The api from message function should take a vumi message, and
-        return a dict with the appropriate values'''
-        channel = yield self.create_channel(
-            self.service, self.redis, TelnetServerTransport, id='channel-id')
-        message = TransportUserMessage.send(
-            content=None, from_addr='+1234', to_addr='+5432',
-            transport_name='testtransport', continue_session=True,
-            helper_metadata={'voice': {}})
-        dct = channel.api_from_message(message)
-        [dct.pop(f) for f in ['timestamp', 'message_id']]
-        self.assertEqual(dct, {
-            'channel_data': {
-                'continue_session': True,
-                'voice': {},
-                },
-            'from': '+1234',
-            'to': '+5432',
-            'channel_id': 'testtransport',
-            'content': None,
-            'reply_to': None,
-            })
-
-    @inlineCallbacks
-    def test_message_from_api(self):
-        yield self.create_channel(
-            self.service, self.redis, TelnetServerTransport, id='channel-id')
-        msg = Channel.message_from_api(
-            'channel-id', {
-                'from': '+1234',
-                'content': None,
-                'channel_data': {
-                    'continue_session': True,
-                    'voice': {},
-                    },
-                })
-        msg = TransportUserMessage.send(**msg)
-        self.assertEqual(msg.get('continue_session'), True)
-        self.assertEqual(msg.get('helper_metadata'), {'voice': {}})
-        self.assertEqual(msg.get('from_addr'), '+1234')
-        self.assertEqual(msg.get('content'), None)
-
-    @inlineCallbacks
-    def test_message_from_api_reply(self):
-        yield self.create_channel(
-            self.service, self.redis, TelnetServerTransport, id='channel-id')
-
-        msg = Channel.message_from_api(
-            'channel-id', {
-                'reply_to': 1234,
-                'content': 'foo',
-                'channel_data': {
-                    'continue_session': True,
-                    'voice': {},
-                },
-            })
-
-        self.assertFalse('to_addr' in msg)
-        self.assertFalse('from_addr' in msg)
-        self.assertEqual(msg['continue_session'], True)
-        self.assertEqual(msg['helper_metadata'], {'voice': {}})
-        self.assertEqual(msg['content'], 'foo')
